Enable colon and set feature
put lib functions in separate utils file
